+++ b4n/ui/views/yaml/view.rs
@@ -3,7 +3,7 @@ use b4n_config::keys::{KeyCombination, KeyCommand};
 use b4n_kube::utils::deserialize_kind;
 use b4n_kube::{ResourceRef, SECRETS};
 use b4n_tasks::commands::{
-    CommandResult, ResourceYamlResult, SetNewResourceYamlOptions, SetResourceYamlAction, SetResourceYamlOptions,
+    CommandResult, ResourceYamlResult, RunPluginOutput, SetNewResourceYamlOptions, SetResourceYamlAction, SetResourceYamlOptions,
 };
 use b4n_tui::widgets::{ActionItem, ActionsListBuilder, Button, CheckBox, Dialog};
 use b4n_tui::{MouseEventKind, ResponseEvent, Responsive, TuiEvent};
@@ -18,7 +18,8 @@ use crate::ui::presentation::{Content, ContentViewer, StyleFallback, StyledLine}
 use crate::ui::views::{View, yaml::YamlContent};
 use crate::ui::widgets::{CommandPalette, FileSelector, Search};
 
-/// YAML view.
+/// YAML view.\
+/// **Note** that it is also used to display plugin command output.
 pub struct YamlView {
     yaml: ContentViewer<YamlContent>,
     app_data: SharedAppData,
@@ -93,6 +94,13 @@ impl YamlView {
         }
     }
 
+    /// Sets header title for the view.
+    pub fn with_title(mut self, title: impl Into<String>, icon: char) -> Self {
+        self.yaml.header.set_title(title);
+        self.yaml.header.set_icon(icon);
+        self
+    }
+
     /// Marks YAML view to switch to edit when data is received.
     pub fn switch_to_edit(&mut self) {
         self.is_edit = true;
@@ -120,7 +128,7 @@ impl YamlView {
             } else if is_current_line {
                 "Line copied to clipboard"
             } else {
-                "YAML content copied to clipboard"
+                "View content copied to clipboard"
             }
         });
     }
@@ -150,15 +158,15 @@ impl YamlView {
             .with_back()
             .with_quit()
             .with_action(
-                ActionItem::action("copy", "copy").with_description("copies YAML to clipboard"),
+                ActionItem::action("copy", "copy").with_description("copies content to clipboard"),
                 Some(KeyCommand::ContentCopy),
             )
             .with_action(
-                ActionItem::action("save", "save").with_description("saves YAML to a file"),
+                ActionItem::action("save", "save").with_description("saves content to a file"),
                 Some(KeyCommand::ContentSave),
             )
             .with_action(
-                ActionItem::action("search", "search").with_description("searches YAML using the provided query"),
+                ActionItem::action("search", "search").with_description("searches content using the provided query"),
                 Some(KeyCommand::SearchOpen),
             );
         if self.yaml.content().is_some_and(Content::is_editable) {
@@ -672,6 +680,19 @@ impl YamlView {
         }
     }
 
+    fn process_new_output(&mut self, result: RunPluginOutput) {
+        let Some(highlighter) = self.worker.borrow().get_highlighter() else {
+            return;
+        };
+        self.yaml.set_content(YamlContent::new(
+            result.styled.into_iter().map(StyledLine::from).collect(),
+            result.output,
+            highlighter,
+            false,
+            StyleFallback::default(),
+        ));
+    }
+
     fn update_view_state(&mut self) {
         if self.state == ViewState::WaitingForClose {
             self.state = ViewState::Closing;
@@ -726,6 +747,9 @@ impl View for YamlView {
                 self.update_view_state();
                 self.footer.show_info(format!("'{name}' YAML saved successfully"), 3_000);
             },
+            CommandResult::RunPluginOutput(Ok(result)) => {
+                self.process_new_output(result);
+            },
             _ => (),
         }
     }
@@ -756,7 +780,7 @@ impl View for YamlView {
         result
     }
 
-    fn draw(&mut self, frame: &mut Frame<'_>, area: Rect) {
+    fn draw(&mut self, frame: &mut Frame<'_>, area: Rect, _has_focus: bool) {
         self.yaml.draw(frame, area, None);
         self.command_palette.draw(frame, frame.area());
         self.search.draw(frame, frame.area());